# laravel/app/Http/Resources/Document/DocumentResource.php
<?php

namespace App\Http\Resources\Document;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

class DocumentResource extends JsonResource
{
    public function toArray(Request $request)
    {
        return parent::toArray($request); // TODO: Change the autogenerated stub
    }
}