# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JBI-3",
        "contact": {
            "email": "admin@mail.ru"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8083/api/",
            "description": "JBI-3"
        }
    ],
    "paths": {
        "/api/v1/admin/document-categories": {
            "get": {
                "tags": [
                    "Admin/DocumentCategory"
                ],
                "summary": "Получение списка категорий документов",
                "description": "Возвращает список категории документов",
                "operationId": "7fce863f9380e9fe211de67936c466c7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsCategoryIndexResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin/DocumentCategory"
                ],
                "summary": "Добавление новой категории документов",
                "operationId": "cb63605b5c3f774def7c86e1b2933ede",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewsCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsCategoryResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/document-categories/{id}": {
            "get": {
                "tags": [
                    "Admin/DocumentCategory"
                ],
                "summary": "Получение категории документов",
                "operationId": "0608977d3714f108b6991b78424b0d80",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id категории",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsCategory"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin/DocumentCategory"
                ],
                "summary": "Обновление категории документов",
                "operationId": "b9f4f520e79976dfd771d549f0d0afff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewsCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsCategoryResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/DocumentCategory"
                ],
                "summary": "Удаление категории документов",
                "operationId": "c9f2d382191083ea560bb29b30cd19e0",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/documents": {
            "get": {
                "tags": [
                    "Admin/Document"
                ],
                "summary": "Получение списка документов",
                "description": "Возвращает список документов",
                "operationId": "dea53e3648acd7e589bd37568e1c5bd2",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentIndexResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin/Document"
                ],
                "summary": "Добавление нового документа",
                "operationId": "ee783825ae32eeb325036ac2c5452610",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/documents/{id}": {
            "get": {
                "tags": [
                    "Admin/Document"
                ],
                "summary": "Получение группы документов",
                "operationId": "26adf7a462fa4a88bb2bbd552d4c036d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id документа",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin/Document"
                ],
                "summary": "Обновление группы документов",
                "operationId": "8342e85aa5a6aa29d39f4f30b85293b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DocumentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/Document"
                ],
                "summary": "Удаление документа",
                "operationId": "9d1f8b8614877e52e1d97f90a3699d11",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResource"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Document": {
                "title": "Document",
                "description": "Модель документа",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Название документа",
                        "type": "string",
                        "example": "JBI"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "group_id": {
                        "title": "group_id",
                        "description": "group_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Активность файла",
                        "type": "boolean",
                        "example": true
                    },
                    "date": {
                        "title": "date",
                        "description": "Дата файла",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-12"
                    },
                    "file": {
                        "title": "file",
                        "description": "file",
                        "type": "string",
                        "example": "[]"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Document"
                }
            },
...
            "NewsCategory": {
                "title": "NewsCategory",
                "description": "Модель категорий новостей",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Название категории",
                        "type": "string",
                        "example": "интересное"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NewsCategory"
                }
            },
...
            "DocumentRequest": {
                "title": "DocumentRequest",
                "description": "Тело запроса",
                "required": [
                    "title",
                    "is_active"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Название документа",
                        "type": "string",
                        "example": "JBI"
                    },
                    "category_id": {
                        "title": "category_id",
                        "description": "category_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "group_id": {
                        "title": "group_id",
                        "description": "group_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "Активность файла",
                        "type": "boolean",
                        "example": true
                    },
                    "date": {
                        "title": "date",
                        "description": "Дата файла",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-12"
                    },
                    "file": {
                        "title": "file",
                        "description": "file",
                        "type": "string",
                        "example": "[]"
                    }
                },
                "type": "object"
            },
...
            "NewsCategoryRequest": {
                "title": "NewsCategoryRequest",
                "description": "Тело запроса",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Название категории",
                        "type": "string",
                        "example": "Акции"
                    }
                },
                "type": "object"
            },
...
            "DeleteAnswerResource": {
                "title": "DeleteAnswerResource",
                "description": "DeleteAnswerResource",
                "properties": {
                    "success": {
                        "title": "success",
                        "description": "success",
                        "type": "integer",
                        "format": "bool",
                        "example": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DeleteAnswerResource"
                }
            },
            "DeleteResource": {
                "title": "DeleteResource",
                "description": "Delete Resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/DeleteAnswerResource"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DeleteResource"
                }
            },
            "DocumentIndexResource": {
                "title": "DocumentIndexResource",
                "description": "DocumentIndex resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Document"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DocumentIndexResource"
                }
            },
            "DocumentResource": {
                "title": "DocumentResource",
                "description": "Document resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Document"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DocumentResource"
                }
            },
...
            "NewsCategoryIndexResource": {
                "title": "NewsCategoryIndexResource",
                "description": "NewsCategoryIndexResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NewsCategory"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NewsCategoryIndexResource"
                }
            },
...
            "NewsCategoryResource": {
                "title": "NewsCategoryResource",
                "description": "News Category Resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "data"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NewsCategoryResource"
                }
            },
...
        }
    }
}